requirements.txt
BigQuery-Python==1.12.0
cachetools==2.0.0
click==6.7
Flask==0.12.2
Flask-API==0.7.1
google-api-python-client==1.6.2
google-auth==1.0.2
google-auth-httplib2==0.0.2
google-cloud-bigquery==0.25.0
google-cloud-core==0.25.0
googleapis-common-protos==1.5.2
httplib2==0.10.3
itsdangerous==0.24
oauth2client==4.1.2
protobuf==3.3.0
pyasn1==0.3.1
pyasn1-modules==0.0.10
python-dateutil==2.6.1
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
Werkzeug==0.12.2
